requirements.txt
spacy==3.1.3
newspaper3k==0.2.8
sumy==0.8.1
pandas==1.3.3
